deploy_action/requirements.txt
requests
PyYaml